refactoring: changed store folder structure
