+++ src/index.js
@@ -9,8 +9,8 @@ import createSagaMiddleware from 'redux-saga';
 import { createBrowserHistory } from 'history';
 import { createReduxHistoryContext } from "redux-first-history";
 import { HistoryRouter } from "redux-first-history/rr6";
-import { rootReducer } from './redux/rootReducer';
-import rootSaga from './redux/sagas/rootSaga';
+import rootReducer from './store/reducer'
+import saga from './store/saga';
 
 const sagaMiddleware = createSagaMiddleware();
 
@@ -28,7 +28,7 @@ export const store = createStore(rootReducer(routerReducer), compose(
 ))
 
 export const history = createReduxHistory(store);
-sagaMiddleware.run(rootSaga);
+sagaMiddleware.run(saga);
 
 ReactDOM.render(
   <React.StrictMode>
