+++ src/pages/PersonDetailsPage/PersonDetailsPage.jsx
@@ -3,7 +3,8 @@ import styles from './PersonDetailsPage.module.scss';
 import { useParams, useNavigate } from 'react-router-dom';
 import { useEffect } from 'react/cjs/react.development';
 import { useDispatch } from 'react-redux';
-import { deletePersonMiddleware, getPersonDetails, showPopup } from '../../redux/actions';
+import { getPersonDetails } from '../../store/personDetailsPage/actions';
+import { deletePersonMiddleware } from '../../store/peopleDB/actions';
 import { useSelector } from 'react-redux';
 import DeletedStamp from '../../components/DeletedStamp/DeletedStamp';
 import ImagePlaceholder from '../../components/ImagePlaceholder/ImagePlaceholder';
