--- src/redux/actions.js
@@ -1,148 +1,0 @@
-import {
-  ADD_LOADER_REQUEST,
-  DECREASE_LOADER_REQUEST,
-  INIT_LIST_OF_PEOPLE,
-  CLEAR_LIST_OF_PEOPLE,
-  SET_USER,
-  DELETE_PERSON,
-  DELETE_PERSON_MIDDLEWARE,
-  CREATE_USER,
-  AUTH_USER,
-  USER_WAS_NOT_FOUND,
-  RESET_USER_WAS_NOT_FOUND_FIELD,
-  ADD_NEW_PERSON,
-  REMOVE_USER,
-  ADD_PEOPLE_ON_PAGE,
-  ADD_COLORS_SET,
-  GET_PERSON_DETAILS,
-  LOAD_COLOR_DETAILS_MIDDLEWARE,
-  SET_COLOR_DETAILS,
-} from './types';
-
-export const onLoader = () => {
-  return {
-    type: ADD_LOADER_REQUEST,
-  }
-}
-
-export const offLoader = () => {
-  return {
-    type: DECREASE_LOADER_REQUEST,
-  }
-}
-
-export const initPeopleCollection = (payload) => {
-  return {
-    type: INIT_LIST_OF_PEOPLE,
-    payload
-  }
-}
-
-export const clearPeopleCollection = () => {
-  return {
-    type: CLEAR_LIST_OF_PEOPLE,
-  }
-}
-
-export const authUser = (userCredentials) => {
-  return {
-    type: AUTH_USER,
-    payload: userCredentials
-  }
-}
-
-export const userWasNotFound = () => {
-  return {
-    type: USER_WAS_NOT_FOUND
-  }
-}
-
-export const resetUserWasNotFoundField = () => {
-  return {
-    type: RESET_USER_WAS_NOT_FOUND_FIELD
-  }
-}
-
-export const createUser = (userCredentials) => {
-  return {
-    type: CREATE_USER,
-    payload: userCredentials
-  }
-}
-
-export const setUser = (user) => {
-  if (!sessionStorage.getItem('token')) {
-    sessionStorage.setItem('token', JSON.stringify(user.token));
-    delete user.token;
-    sessionStorage.setItem('user', JSON.stringify(user));
-  }
-
-  return {
-    type: SET_USER,
-    payload: user
-  }
-}
-
-export const removeUser = () => {
-  sessionStorage.removeItem('user');
-  sessionStorage.removeItem('token');
-  return {
-    type: REMOVE_USER
-  }
-}
-
-export const addNewPerson = (payload) => {
-  return {
-    type: ADD_NEW_PERSON,
-    payload
-  }
-}
-
-export const deletePerson = (personId) => {
-  return {
-    type: DELETE_PERSON,
-    payload: personId
-  }
-}
-
-export const loadPeoplePage = (people) => {
-  return {
-    type: ADD_PEOPLE_ON_PAGE,
-    payload: people
-  }
-}
-
-export const addColorsOnPage = (colors) => {
-  return {
-    type: ADD_COLORS_SET,
-    payload: colors
-  }
-}
-
-export const getPersonDetails = (details) => {
-  return {
-    type: GET_PERSON_DETAILS,
-    payload: details
-  }
-}
-
-export const deletePersonMiddleware = (id) => {
-  return {
-    type: DELETE_PERSON_MIDDLEWARE,
-    payload: id
-  }
-}
-
-export const loadColorDetailsMiddleware = (id) => {
-  return {
-    type: LOAD_COLOR_DETAILS_MIDDLEWARE,
-    payload: id
-  }
-}
-
-export const setColorDetails = (fullInfo) => {
-  return {
-    type: SET_COLOR_DETAILS,
-    payload: fullInfo
-  }
-}
