--- src/redux/types.js
@@ -1,26 +1,0 @@
-export const ADD_LOADER_REQUEST = 'APP/ADD_LOADER_REQUEST';
-export const DECREASE_LOADER_REQUEST = 'APP/DECREASE_LOADER_REQUEST';
-
-export const INIT_LIST_OF_PEOPLE = 'PEOPLE/INIT_LIST_OF_PEOPLE';
-export const CLEAR_LIST_OF_PEOPLE = 'PEOPLE/CLEAR_LIST_OF_PEOPLE';
-export const ADD_NEW_PERSON = 'PEOPLE/ADD_NEW_PERSON';
-export const LOGOUT_USER = 'PEOPLE/LOGOUT_USER';
-export const DELETE_PERSON = 'PEOPLE/DELETE_PERSON';
-export const DELETE_PERSON_MIDDLEWARE = 'PEOPLE/DELETE_PERSON_MIDDLEWARE'
-
-export const CREATE_USER = 'USER/CREATE_USER';
-export const AUTH_USER = 'USER/AUTH_USER';
-export const SET_USER = 'USER/SET_USER';
-export const REMOVE_USER = 'USER/REMOVE_USER';
-export const USER_WAS_NOT_FOUND = 'USER/USER_WAS_NOT_FOUND';
-export const RESET_USER_WAS_NOT_FOUND_FIELD = 'USER/RESET_USER_WAS_NOT_FOUND_FIELD';
-
-export const LOAD_PEOPLE_SET = 'PEOPLE_PAGE/LOAD_PEOPLE_SET';
-export const ADD_PEOPLE_ON_PAGE = 'PEOPLE_PAGE/ADD_PEOPLE_ON_PAGE';
-
-export const ADD_COLORS_SET = 'COLORS_PAGE/ADD_COLORS_SET';
-
-export const GET_PERSON_DETAILS = 'PERSON_PAGE/GET_PERSON_DETAILS';
-
-export const LOAD_COLOR_DETAILS_MIDDLEWARE = 'COLOR_PAGE/LOAD_COLOR_DETAILS_MIDDLEWARE';
-export const SET_COLOR_DETAILS = 'COLOR_PAGE/SET_COLOR_DETAILS';
