+++ src/store/app/sagas/setMainData.js
@@ -1,8 +1,8 @@
 import { call } from 'redux-saga/effects';
-import setPeopleCollection from './setPeopleCollection';
-import setUserToState from './setUser';
+import setPeopleCollection from '../../peopleDB/sagas/setPeopleCollection';
+import setUserToState from '../../user/sagas/setUserToState';
 
-export default function* setMainDataSaga() {
+export default function* setMainData() {
   const user = yield JSON.parse(sessionStorage.getItem('user'));
 
   if (user) {
