+++ src/store/colorsPage/actions.js
@@ -1,0 +1,8 @@
+import { ADD_COLORS_SET } from '../types';
+
+export const addColorsOnPage = (colors) => {
+  return {
+    type: ADD_COLORS_SET,
+    payload: colors
+  }
+}
