+++ src/store/personDetailsPage/actions.js
@@ -1,0 +1,8 @@
+import { GET_PERSON_DETAILS } from '../types';
+
+export const getPersonDetails = (details) => {
+  return {
+    type: GET_PERSON_DETAILS,
+    payload: details
+  }
+}
